requirements.txt
mysql-connector==2.2.9
qrl==4.0.1
requests==2.26.0
